Moved Dashboards to their own components Created Home components for both Customer/ServiceProvider Updated logic to show Home component on dash when no other components are loaded using: {props.children ? props.children : <Home />}
***IMPORTANT***
Auth on Service provider dash currently allows Customer Role as Service provider sign up
is not working. Make sure to update this when service provider sign up is working again.
