+++ frontend/pages/Customer/NewRequest/index.js
@@ -1,9 +1,7 @@
 // import statements
 import * as React from "react";
 import NewRequest from "../../../components/Requests/NewRequest";
-import CustomerDash from "../index";
-
-
+import CustomerDash from "../../../components/Dashboard/CustomerDashboard";
 
 export default function NewRequestScreen() {
   return (
