+++ frontend/pages/Customer/index.js
@@ -1,300 +1,7 @@
 // import statements
 import * as React from "react";
-import { styled, createTheme, ThemeProvider } from "@mui/material/styles";
-import CssBaseline from "@mui/material/CssBaseline";
-import MuiDrawer from "@mui/material/Drawer";
-import Box from "@mui/material/Box";
-import MuiAppBar from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import List from "@mui/material/List";
-import Typography from "@mui/material/Typography";
-import Divider from "@mui/material/Divider";
-``;
-import IconButton from "@mui/material/IconButton";
-import Badge from "@mui/material/Badge";
-import Container from "@mui/material/Container";
-import Grid from "@mui/material/Grid";
-import Paper from "@mui/material/Paper";
-//import Link from "@mui/material/Link";
-import Link from "next/link";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import LogoutIcon from '@mui/icons-material/Logout';
+import CustomerDash from "../../components/Dashboard/CustomerDashboard";
 
-//import Account from "../../../components/Account";
-import AccountCircleRoundedIcon from "@mui/icons-material/AccountCircleRounded";
-import InboxIcon from "@mui/icons-material/MoveToInbox";
-import ExpandLess from "@mui/icons-material/ExpandLess";
-import ExpandMore from "@mui/icons-material/ExpandMore";
-import StarBorder from "@mui/icons-material/StarBorder";
-import Collapse from "@mui/material/Collapse";
-
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import ListSubheader from "@mui/material/ListSubheader";
-import DashboardIcon from "@mui/icons-material/Dashboard";
-import ShoppingCartIcon from "@mui/icons-material/ShoppingCart";
-import PeopleIcon from "@mui/icons-material/People";
-import BarChartIcon from "@mui/icons-material/BarChart";
-import LayersIcon from "@mui/icons-material/Layers";
-import AssignmentIcon from "@mui/icons-material/Assignment";
-import { useCurrentUser } from "../../components/hooks/CurrentUserContext";
-import withAuth from "../../components/router/withAuth";
-import HomeIcon from "@mui/icons-material/Home";
-import { blue } from "@mui/material/colors";
-
-const drawerWidth = 240;
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(["width", "margin"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(["width", "margin"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  "& .MuiDrawer-paper": {
-    position: "relative",
-    whiteSpace: "nowrap",
-    width: drawerWidth,
-    transition: theme.transitions.create("width", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    boxSizing: "border-box",
-    ...(!open && {
-      overflowX: "hidden",
-      transition: theme.transitions.create("width", {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-      }),
-      width: theme.spacing(7),
-      [theme.breakpoints.up("sm")]: {
-        width: theme.spacing(9),
-      },
-    }),
-  },
-}));
-
-const mdTheme = createTheme();
-
-function CustomerDash(props) {
-  const [open, setOpen] = React.useState(true);
-  const toggleDrawer = () => {
-    setOpen(!open);
-  };
-
-  const [newOpen, setnewOpen] = React.useState(true);
-
-  const handleClick = () => {
-    setnewOpen(!newOpen);
-  };
-
-  return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: "flex" }}>
-        <CssBaseline />
-        <AppBar position="absolute" open={open}>
-          <Toolbar
-            sx={{
-              pr: "24px", // keep right padding when drawer closed
-            }}
-          >
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={toggleDrawer}
-              sx={{
-                marginRight: "36px",
-                ...(open && { display: "none" }),
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography
-              component="h1"
-              variant="h6"
-              color="inherit"
-              noWrap
-              sx={{ flexGrow: 1 }}
-            >
-              Dashboard
-            </Typography>
-            <IconButton color="inherit">
-              <Link href="/Customer/account" color="inherit">
-                {
-                  <AccountCircleRoundedIcon
-                    style={{ color: "#FFFFFF" }}
-                    fontSize="large"
-                  />
-                }
-              </Link>
-            </IconButton>
-            <IconButton color="inherit">
-              {/* TODO: add logout functionality to this link */}
-              <Link href="/" color="inherit">
-                {
-                  <LogoutIcon
-                    style={{ color: "#FFFFFF" }}
-                    fontSize="large"
-                  />
-                }
-              </Link>
-            </IconButton>
-          </Toolbar>
-        </AppBar>
-        <Drawer variant="permanent" open={open}>
-          <Toolbar
-            sx={{
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "flex-end",
-              px: [1],
-            }}
-          >
-            <IconButton onClick={toggleDrawer}>
-              <ChevronLeftIcon />
-            </IconButton>
-          </Toolbar>
-          <Divider />
-          <List component="nav">
-            <React.Fragment>
-              <Link href="./" passHref legacyBehavior color="inherit">
-                <ListItemButton>
-                  <ListItemIcon>
-                    <HomeIcon />
-                  </ListItemIcon>
-                  <ListItemText primary="Home" />
-                </ListItemButton>
-              </Link>
-              <Divider sx={{ my: 1 }} />
-              <ListItemButton onClick={handleClick}>
-                <ListItemIcon>
-                  <InboxIcon />
-                </ListItemIcon>
-                <ListItemText primary="Requests" />
-                {newOpen ? <ExpandLess /> : <ExpandMore />}
-              </ListItemButton>
-              <Collapse in={newOpen} timeout="auto" unmountOnExit>
-                <List component="div" disablePadding>
-                  <Link
-                    href="/Customer/NewRequest"
-                    passHref
-                    legacyBehavior
-                    color="inherit"
-                  >
-                    <ListItemButton sx={{ pl: 4 }}>
-                      <ListItemIcon>
-                        <StarBorder />
-                      </ListItemIcon>
-                      <ListItemText primary="New Request" />
-                    </ListItemButton>
-                  </Link>
-                </List>
-              </Collapse>
-              <Collapse in={newOpen} timeout="auto" unmountOnExit>
-                <List component="div" disablePadding>
-                  <Link
-                    href="/Customer/CurrentRequest"
-                    passHref
-                    legacyBehavior
-                    color="inherit"
-                  >
-                    <ListItemButton sx={{ pl: 4 }}>
-                      <ListItemIcon>
-                        <StarBorder />
-                      </ListItemIcon>
-                      <ListItemText primary="Current Requests" />
-                    </ListItemButton>
-                  </Link>
-                </List>
-              </Collapse>
-              <Collapse in={newOpen} timeout="auto" unmountOnExit>
-                <List component="div" disablePadding>
-                  <Link
-                    href="/Customer/PastRequests"
-                    passHref
-                    legacyBehavior
-                    color="inherit"
-                  >
-                    <ListItemButton sx={{ pl: 4 }}>
-                      <ListItemIcon>
-                        <StarBorder />
-                      </ListItemIcon>
-                      <ListItemText primary="Past Requests" />
-                    </ListItemButton>
-                  </Link>
-                </List>
-              </Collapse>
-              <Link
-                href="/Customer/account"
-                passHref
-                legacyBehavior
-                color="inherit"
-              >
-                <ListItemButton>
-                  <ListItemIcon>
-                    <AccountCircleRoundedIcon />
-                  </ListItemIcon>
-                  <ListItemText primary="Account Details" />
-                </ListItemButton>
-              </Link>
-              <Link
-                href="/Customer/Report"
-                passHref
-                legacyBehavior
-                color="inherit"
-              >
-                <ListItemButton>
-                  <ListItemIcon>
-                    <BarChartIcon />
-                  </ListItemIcon>
-                  <ListItemText primary="Reports" />
-                </ListItemButton>
-              </Link>
-            </React.Fragment>
-            <Divider sx={{ my: 1 }} />
-          </List>
-        </Drawer>
-        <Box
-          component="main"
-          sx={{
-            backgroundColor: (theme) =>
-              theme.palette.mode === "light"
-                ? theme.palette.grey[100]
-                : theme.palette.grey[900],
-            flexGrow: 1,
-            height: "100vh",
-            overflow: "auto",
-          }}
-        >
-          <Toolbar />
-          <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-            {props.children}
-          </Container>
-        </Box>
-      </Box>
-    </ThemeProvider>
-  );
+export default function Dashboard() {
+  return <CustomerDash />;
 }
-
-export default withAuth(CustomerDash, ["ROLE_CUSTOMER"]);
-// export default function Dashboard() {
-//   return <CustomerDash />;
-// }
